requirements.txt
# Flask for the API
Flask

# Core LangChain
langchain
langchain-core==0.3.40
langchain-community

# Google Gemini
google-generativeai==0.8.5
langchain-google-genai==2.0.10
google-ai-generativelanguage==0.6.15

# Database
pyodbc
sqlalchemy
pydantic

# Helper libraries
pandas
python-dotenv
ipykernel
ipywidgets